+++ src/app/admin/products/components/product-list-component/product-list-component.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService } from '../../../../products/services/products.service';
-import { CartService } from '../../../../cart/services/cart.service';
 import { Product } from '../../models/product.model';
 
 @Component({
@@ -13,7 +12,7 @@ export class ProductListComponentComponent implements OnInit {
   products: Array<Product>;
   // products: Promise<Product[]>;
 
-  constructor(public productsService: ProductsService, private cartService: CartService) { }
+  constructor(public productsService: ProductsService) { }
 
   ngOnInit() {
     this.getProducts();