+++ src/app/products/services/products.service.ts
@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Product } from '../models/product.model';
-import { Category } from '../../shared/enums/enums';
 import { LocalStorageService } from '../../core/services/local-storage.service';
-import { identifierModuleUrl } from '@angular/compiler';
 
 @Injectable({
   providedIn: 'root'