+++ changelog.md
@@ -48,3 +48,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Routing for the application, admin module, products, cart
 - AuthGuard, CanLoad, LoadChildren, Reacting to Router Events (Title), Meta Service
 - LocalStorage to the core services
+
+## [5.0.0] - 2019-03-29
+### Added
+- JSON-server (db/db.json)
+- AppSettingsService to save settings in LocalStorage
+- httpInterceptorProviders to get request duration
+- AutoUnsubscribe decorator
+### Changed
+- redid CartService to work with http according to the Observable scheme
+- redid ProductsService to work with http according to the Promise scheme